refactor(dashboard-tsr): split largest table/insights components (#1392)
Extract co-located subcomponents and a shared query-table from the
biggest faithful-port files. Behavior-preserving.

- expensive-queries-table 1133 -> 379 (table/* + query-tables/* shared)
- insights route 938 -> 51 (-insights/* stat components)
- running-queries-table 1014 -> slim (table/* incl. use-kill-query)
- assistant-ui thread: extract format helpers
